+++ internal/interface/cli/root.go
@@ -71,6 +71,7 @@ func NewRootCommand() *cobra.Command {
 	root.AddCommand(newLogoutCommand())
 	root.AddCommand(newWhoamiCommand())
 	root.AddCommand(newSandboxCommand())
+	root.AddCommand(newToolsCommand())
 	root.AddCommand(newModelsCommand())
 	return root
 }